LayoutTests/inspector/dom-debugger/event-listener-breakpoints-all-events-expected.txt

CONSOLE MESSAGE: BREAKPOINT ACTION LOG 1
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 2
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 3
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 4
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 1
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 2
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 3
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 4
Tests for Event Listener breakpoints.


== Running test suite: DOMDebugger.Event.Listener
-- Running test case: DOMDebugger.Event.Listener.AllEventsBreakpoint
Creating "listener" Event Breakpoint...
Adding "listener" Event Breakpoint...
Firing "click" on body...
PASS: Should pause before event handler is run.
CALL STACK:
0: [F] handleBody_click
1: [F] trigger_click
2: [P] Global Code
Firing "custom" on body...
PASS: Should pause before event handler is run.
CALL STACK:
0: [F] handleBody_custom
1: [F] trigger_custom
2: [P] Global Code
PASS: Should have paused twice.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Listener.AllEventsBreakpoint.BuiltinOptions.Condition
Creating "listener" Event Breakpoint...
Adding "listener" Event Breakpoint...

Setting condition to 'false'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Setting condition to 'true'...

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: DOMDebugger.Event.Listener.AllEventsBreakpoint.BuiltinOptions.Condition.ConsoleCommandLineAPI
Adding saved console value 'false'...
Creating "listener" Event Breakpoint...
Adding "listener" Event Breakpoint...

Setting condition to saved console value...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Adding saved console value 'true'...
Setting condition to saved console value...

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: DOMDebugger.Event.Listener.AllEventsBreakpoint.BuiltinOptions.IgnoreCount
Creating "listener" Event Breakpoint...
Adding "listener" Event Breakpoint...

Setting ignoreCount to '2'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: DOMDebugger.Event.Listener.AllEventsBreakpoint.BuiltinOptions.Action.Log
Creating "listener" Event Breakpoint...
Adding "listener" Event Breakpoint...

Adding log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: DOMDebugger.Event.Listener.AllEventsBreakpoint.BuiltinOptions.Actions.Evaluate
Creating "listener" Event Breakpoint...
Adding "listener" Event Breakpoint...

Adding evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: DOMDebugger.Event.Listener.AllEventsBreakpoint.BuiltinOptions.Actions.Evaluate.ConsoleCommandLineAPI
Adding saved console value '1'...
Creating "listener" Event Breakpoint...
Adding "listener" Event Breakpoint...

Adding evaluate action using saved console value...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Adding saved console value '2'...
Editing evaluate action using saved console value...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Adding saved console value '3'...
Editing evaluate action using saved console value...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Adding saved console value '4'...
Editing evaluate action using saved console value...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: DOMDebugger.Event.Listener.AllEventsBreakpoint.CustomOptions.Condition
Creating "listener" Event Breakpoint...
Adding "listener" Event Breakpoint...

Setting condition to 'false'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Setting condition to 'true'...

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: DOMDebugger.Event.Listener.AllEventsBreakpoint.CustomOptions.Condition.ConsoleCommandLineAPI
Adding saved console value 'false'...
Creating "listener" Event Breakpoint...
Adding "listener" Event Breakpoint...

Setting condition to saved console value...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Adding saved console value 'true'...
Setting condition to saved console value...

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: DOMDebugger.Event.Listener.AllEventsBreakpoint.CustomOptions.IgnoreCount
Creating "listener" Event Breakpoint...
Adding "listener" Event Breakpoint...

Setting ignoreCount to '2'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: DOMDebugger.Event.Listener.AllEventsBreakpoint.CustomOptions.Action.Log
Creating "listener" Event Breakpoint...
Adding "listener" Event Breakpoint...

Adding log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: DOMDebugger.Event.Listener.AllEventsBreakpoint.CustomOptions.Actions.Evaluate
Creating "listener" Event Breakpoint...
Adding "listener" Event Breakpoint...

Adding evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: DOMDebugger.Event.Listener.AllEventsBreakpoint.CustomOptions.Actions.Evaluate.ConsoleCommandLineAPI
Adding saved console value '1'...
Creating "listener" Event Breakpoint...
Adding "listener" Event Breakpoint...

Adding evaluate action using saved console value...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Adding saved console value '2'...
Editing evaluate action using saved console value...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Adding saved console value '3'...
Editing evaluate action using saved console value...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Adding saved console value '4'...
Editing evaluate action using saved console value...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.
